Enable bitnami password rotation job
For some reason we're changing the pgpassword at every deploy, which
isn't ideal. I'm not sure how to turn it off so for now let's try
enabling the password rotation script from the postgres helm chart

I'm trying to configure a dependency helm chart here even though it's
not in the Values.yaml file from the authentik helm chart. Maybe this
will work?